refactor: split IaC code from ScanCollection to IaCScanCollection
